Move github specific methods on User into seperate modules
